requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.13
django-cors-headers==3.12.0
django-push-notifications==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.7
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2024.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
